Разбил классы кнопок по разным файлам
